"Fix" the summarize_tree.c file
For some reason, I'm assuming because my personal computer is using ubuntu, our solution for this one doesn't work. It seems like it is because the . and .. appear at the end of the readdir list instead of the beginnning. So I wrote some more "generic" code that seems to work. However the second part of the && in the conditional might not be necessary, and I don't know if I like it because it will ignore files like .gitignore.